+++ nucypher/blockchain/eth/providers.py
@@ -14,6 +14,8 @@ GNU Affero General Public License for more details.
 You should have received a copy of the GNU Affero General Public License
 along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
 """
+
+
 from typing import Union
 from urllib.parse import urlparse
 
@@ -23,7 +25,7 @@ from web3 import HTTPProvider, IPCProvider, WebsocketProvider
 from web3.providers import BaseProvider
 from web3.providers.eth_tester.main import EthereumTesterProvider
 
-from nucypher.blockchain.eth.clients import NuCypherGethDevProcess
+from nucypher.blockchain.eth.processes import NuCypherGethDevProcess
 from nucypher.exceptions import DevelopmentInstallationRequired
